\name{pacourse-package}
\alias{pacourse-package}
\alias{pacourse}
\docType{package}
\title{
\packageTitle{pacourse}
}
\description{
\packageDescription{pacourse}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pacourse}
\packageIndices{pacourse}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{pacourse}

Maintainer: \packageMaintainer{pacourse}
}
\references{
~~ Literature or other references for background information ~~
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
